Adds sensible restrictions to bancounts
Pre-amble to letting users ban multiple things. They now can't ban more
than 3 words, and can't delete all of their banned words.

+++ bannedWordServer/routes/banroute.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
 
-from bannedWordServer.constants.errors import DuplicateResourceError, NotFoundError, InvalidTypeError, AuthenticationError
+from bannedWordServer.constants.errors import ValidationError, DuplicateResourceError, NotFoundError, InvalidTypeError, AuthenticationError
 from bannedWordServer.models.ban import Ban
 from bannedWordServer.models.server import Server
 from bannedWordServer.routes.resource import Resource
@@ -40,23 +40,31 @@ class BanRoute(Resource):
 		if not isinstance(banned_word, str): raise InvalidTypeError
 		server_to_modify = session.query(Server).filter_by(server_id=serverid).first()
 		if not server_to_modify: raise NotFoundError
+		if len(server_to_modify.banned_words) == 3:
+			raise ValidationError
 
 		already_exists = session.query(Ban).filter_by(server_id=serverid, banned_word=banned_word).first()
 		if already_exists: raise DuplicateResourceError
 
-		new_ban = Ban(server_id=serverid, banned_word=banned_word)
+		new_ban = Ban(server_id=serverid,
+			banned_word=banned_word,
+			infracted_at=datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+			calledout_at=(datetime.now() - timedelta(weeks=52)).strftime("%Y-%m-%d %H:%M:%S"))
 		session.add(new_ban)
 
 		return session.query(Ban).filter_by(server_id=serverid, banned_word=banned_word).first().to_dict()
 
-	def post_one(self, session, authToken, banid: str, banned_word: str) -> dict:
+	def post_one(self, session, authToken, serverid: str, banid: str, banned_word: str) -> dict:
 		if not authenticateBotOnly(authToken): raise AuthenticationError
 		try:
+			serverid = int(serverid)
 			banid = int(banid)
 		except:
 			raise InvalidTypeError
 		if not isinstance(banned_word, str): raise InvalidTypeError
-		ban = session.query(Ban).filter_by(rowid=banid).first()
+		server_to_modify = session.query(Server).filter_by(server_id=serverid).first()
+		if not server_to_modify: raise NotFoundError
+		ban = session.query(Ban).filter_by(server_id=serverid, rowid=banid).first()
 		if not ban:	raise NotFoundError
 
 		ban.banned_word = banned_word
@@ -64,13 +72,21 @@ class BanRoute(Resource):
 		ban.calledout_at = (datetime.now() - timedelta(weeks=52)).strftime("%Y-%m-%d %H:%M:%S")
 		return self.get_one(session, authToken, banid)
 
-	def delete(self, session, authToken, banid: str):
+	def delete(self, session, authToken, serverid: str, banid: str):
 		if not authenticateBotOnly(authToken): raise AuthenticationError
 		try:
 			banid = int(banid)
+			serverid = int(serverid)
 		except:
 			raise InvalidTypeError
-		ban = session.query(Ban).filter_by(rowid=banid).first()
+
+		server_to_modify = session.query(Server).filter_by(server_id=serverid).first()
+		if not server_to_modify: raise NotFoundError
+		if len(server_to_modify.banned_words) == 1:
+			raise ValidationError
+
+		ban = session.query(Ban).filter_by(server_id=serverid, rowid=banid).first()
+
 		if not ban:	raise NotFoundError
 		session.delete(ban)
